+++ models/structures.go
@@ -4,10 +4,14 @@ import (
 	"os"
 )
 
+// Utilisateurs lus dans .beeusers.conf
 type User struct {
 	Password string `toml:"password"`
 	IsAdmin  bool   `toml:"is_admin"`
 	IsEditor bool   `toml:"is_editor"`
+}
+type Access struct {
+	IsEditor bool `toml:"is_editor"`
 }
 
 // BeeConfig structure du fichier de configuration de l'application app.conf
@@ -34,10 +38,6 @@ type BeeConfig struct {
 	Index       []byte              // index des mots full text
 	Users       map[string]User     // Les utilisateurs
 	BeeFiles    map[string]*BeeFile // index des beefiles
-	// runmode = dev
-	Runmode  string // environnement d'exécution: dev,production
-	UserDev  string
-	GroupDev string
 }
 
 // Bee Context webapp courante dans la session
@@ -53,10 +53,8 @@ type BeeDir struct {
 	BeeFiles      map[string]*BeeFile // la liste des fichiers de content
 	Keywords      []string            // les hashtags du dossier
 	KeywordsAlbum []string            // les hashtags de l'album
-	Users         map[string]Access
-}
-type Access struct {
-	IsEditor bool `toml:"is_editor"`
+	Users         map[string]Access   // droit d'accès des utilisateurs
+	Couverture    string              // id de l'image couverture de l'album
 }
 
 // BeeFile propriétés d'un fichier dans le sous-dossier BeeDir
@@ -111,6 +109,7 @@ type BeeFile struct {
 	Latitude     string
 	Longitude    string
 	LensModel    string
+	Source       string // =id pour inquer que l'image est la couverture de l'album
 }
 
 // BeePathInfo as