+++ static/js/bee_script.js
@@ -428,20 +428,18 @@ $(document).ready(function () {
     var $return = $(this).data('return');
     var $form = $('#form_meta_id');
     $form.find("input[name='return']").val($return)
-    if (quillInitilized) {
+    if (quillModified) {
       var $comment = quill.root.innerHTML;
       $form.find("input[name='doc']").val($comment);
-      if (quillModified) {
-        captureElement('.ql-editor')
-          .then(function (canvas) {
-            var dataUrl = canvas.toDataURL('image/png');
-            $form.find("input[name='canvas']").val(dataUrl);
-            $form.submit();
-          })
-          .catch(function (error) {
-            console.error('Error capturing element:', error);
-          });
-      }
+      captureElement('.ql-editor')
+        .then(function (canvas) {
+          var dataUrl = canvas.toDataURL('image/png');
+          $form.find("input[name='canvas']").val(dataUrl);
+          $form.submit();
+        })
+        .catch(function (error) {
+          console.error('Error capturing element:', error);
+        });
     } else {
       $form.submit();
     }